backend/requirements.txt
Flask
flask-cors
Flask-Limiter
WTForms
python-dotenv
requests
